# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Label Studio API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /label-sheets:
    get:
      operationId: getLabelSheets
      tags: [labelSheets]
      summary: List all label sheet templates
      responses:
        "200":
          description: List of label sheet templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LabelSheet"
    post:
      operationId: createLabelSheet
      tags: [labelSheets]
      summary: Create a custom label sheet template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLabelSheetBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabelSheet"
  /label-sheets/{id}:
    get:
      operationId: getLabelSheet
      tags: [labelSheets]
      summary: Get a label sheet template by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Label sheet template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabelSheet"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateLabelSheet
      tags: [labelSheets]
      summary: Update a label sheet template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLabelSheetBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabelSheet"
    delete:
      operationId: deleteLabelSheet
      tags: [labelSheets]
      summary: Delete a label sheet template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /label-templates:
    get:
      operationId: getLabelTemplates
      tags: [labelTemplates]
      summary: List all label templates
      responses:
        "200":
          description: List of label templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LabelTemplate"
    post:
      operationId: createLabelTemplate
      tags: [labelTemplates]
      summary: Create a label template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLabelTemplateBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabelTemplate"
  /label-templates/{id}:
    get:
      operationId: getLabelTemplate
      tags: [labelTemplates]
      summary: Get a label template by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Label template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabelTemplate"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateLabelTemplate
      tags: [labelTemplates]
      summary: Update a label template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLabelTemplateBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabelTemplate"
    delete:
      operationId: deleteLabelTemplate
      tags: [labelTemplates]
      summary: Delete a label template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /products:
    get:
      operationId: getProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: productType
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /print-jobs:
    get:
      operationId: getPrintJobs
      tags: [printJobs]
      summary: List all print jobs
      responses:
        "200":
          description: List of print jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PrintJob"
    post:
      operationId: createPrintJob
      tags: [printJobs]
      summary: Create a new print job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePrintJobBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrintJob"
  /print-jobs/{id}:
    get:
      operationId: getPrintJob
      tags: [printJobs]
      summary: Get a print job by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Print job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrintJob"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updatePrintJob
      tags: [printJobs]
      summary: Update a print job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePrintJobBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrintJob"
    delete:
      operationId: deletePrintJob
      tags: [printJobs]
      summary: Delete a print job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    LabelSheet:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        brand:
          type: string
        code:
          type: string
        pageWidth:
          type: number
        pageHeight:
          type: number
        labelWidth:
          type: number
        labelHeight:
          type: number
        labelsAcross:
          type: integer
        labelsDown:
          type: integer
        topMargin:
          type: number
        leftMargin:
          type: number
        horizontalGap:
          type: number
        verticalGap:
          type: number
        shape:
          type: string
          enum: [rectangle, circle, oval]
        isCustom:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - brand
        - code
        - pageWidth
        - pageHeight
        - labelWidth
        - labelHeight
        - labelsAcross
        - labelsDown
        - topMargin
        - leftMargin
        - horizontalGap
        - verticalGap
        - shape
        - isCustom
        - createdAt
        - updatedAt
    CreateLabelSheetBody:
      type: object
      properties:
        name:
          type: string
        brand:
          type: string
        code:
          type: string
        pageWidth:
          type: number
        pageHeight:
          type: number
        labelWidth:
          type: number
        labelHeight:
          type: number
        labelsAcross:
          type: integer
        labelsDown:
          type: integer
        topMargin:
          type: number
        leftMargin:
          type: number
        horizontalGap:
          type: number
        verticalGap:
          type: number
        shape:
          type: string
          enum: [rectangle, circle, oval]
      required:
        - name
        - brand
        - code
        - pageWidth
        - pageHeight
        - labelWidth
        - labelHeight
        - labelsAcross
        - labelsDown
        - topMargin
        - leftMargin
        - horizontalGap
        - verticalGap
        - shape
    UpdateLabelSheetBody:
      type: object
      properties:
        name:
          type: string
        brand:
          type: string
        code:
          type: string
        pageWidth:
          type: number
        pageHeight:
          type: number
        labelWidth:
          type: number
        labelHeight:
          type: number
        labelsAcross:
          type: integer
        labelsDown:
          type: integer
        topMargin:
          type: number
        leftMargin:
          type: number
        horizontalGap:
          type: number
        verticalGap:
          type: number
        shape:
          type: string
          enum: [rectangle, circle, oval]
    LabelTemplate:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        labelSheetId:
          type: integer
          nullable: true
        zones:
          type: object
        previewNotes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - zones
        - createdAt
        - updatedAt
    CreateLabelTemplateBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        labelSheetId:
          type: integer
        zones:
          type: object
        previewNotes:
          type: string
      required:
        - name
        - zones
    UpdateLabelTemplateBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        labelSheetId:
          type: integer
          nullable: true
        zones:
          type: object
        previewNotes:
          type: string
...
    Product:
      type: object
      properties:
        id:
          type: integer
        productType:
          type: string
          enum: [soy_candle, room_spray, room_diffuser, other]
        name:
          type: string
        scentName:
          type: string
        scentNotes:
          type: string
          nullable: true
        size:
          type: string
        weight:
          type: string
          nullable: true
        ingredients:
          type: string
          nullable: true
        instructions:
          type: string
          nullable: true
        burnTime:
          type: string
          nullable: true
        waxType:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        warnings:
          type: string
          nullable: true
        sku:
          type: string
          nullable: true
        isActive:
          type: boolean
        labelTemplateId:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - productType
        - name
        - scentName
        - size
        - isActive
        - createdAt
        - updatedAt
    CreateProductBody:
      type: object
      properties:
        productType:
          type: string
          enum: [soy_candle, room_spray, room_diffuser, other]
        name:
          type: string
        scentName:
          type: string
        scentNotes:
          type: string
        size:
          type: string
        weight:
          type: string
        ingredients:
          type: string
        instructions:
          type: string
        burnTime:
          type: string
        waxType:
          type: string
        location:
          type: string
        warnings:
          type: string
        sku:
          type: string
        isActive:
          type: boolean
        labelTemplateId:
          type: integer
      required:
        - productType
        - name
        - scentName
        - size
    UpdateProductBody:
      type: object
      properties:
        productType:
          type: string
          enum: [soy_candle, room_spray, room_diffuser, other]
        name:
          type: string
        scentName:
          type: string
        scentNotes:
          type: string
        size:
          type: string
        weight:
          type: string
        ingredients:
          type: string
        instructions:
          type: string
        burnTime:
          type: string
        waxType:
          type: string
        location:
          type: string
        warnings:
          type: string
        sku:
          type: string
        isActive:
          type: boolean
        labelTemplateId:
          type: integer
          nullable: true
    PrintJobItem:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: integer
        productName:
          type: string
        productScentName:
          type: string
        productType:
          type: string
        productSize:
          type: string
      required:
        - productId
        - quantity
    CreatePrintJobItemBody:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: integer
      required:
        - productId
        - quantity
    PrintJob:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        labelSheetId:
          type: integer
        labelSheetName:
          type: string
        labelSheetCode:
          type: string
        labelsPerSheet:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/PrintJobItem"
        totalLabels:
          type: integer
        totalSheets:
          type: integer
        status:
          type: string
          enum: [draft, ready, printed]
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - labelSheetId
        - labelSheetName
        - labelSheetCode
        - labelsPerSheet
        - items
        - totalLabels
        - totalSheets
        - status
        - createdAt
        - updatedAt
    CreatePrintJobBody:
      type: object
      properties:
        name:
          type: string
        labelSheetId:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/CreatePrintJobItemBody"
        notes:
          type: string
      required:
        - name
        - labelSheetId
        - items
    UpdatePrintJobBody:
      type: object
      properties:
        name:
          type: string
        labelSheetId:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/CreatePrintJobItemBody"
        notes:
          type: string
        status:
          type: string
          enum: [draft, ready, printed]
...