# lib/api-zod/src/generated/types/createLabelTemplateBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Label Studio API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateLabelTemplateBodyZones } from "./createLabelTemplateBodyZones";

export interface CreateLabelTemplateBody {
  name: string;
  description?: string;
  labelSheetId?: number;
  zones: CreateLabelTemplateBodyZones;
  previewNotes?: string;
}