# lib/api-zod/src/generated/types/createPrintJobBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Label Studio API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreatePrintJobItemBody } from "./createPrintJobItemBody";

export interface CreatePrintJobBody {
  name: string;
  labelSheetId: number;
  items: CreatePrintJobItemBody[];
  notes?: string;
}