# lib/api-zod/src/generated/types/createProductBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Label Studio API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateProductBodyProductType } from "./createProductBodyProductType";

export interface CreateProductBody {
  productType: CreateProductBodyProductType;
  name: string;
  scentName: string;
  scentNotes?: string;
  size: string;
  weight?: string;
  ingredients?: string;
  instructions?: string;
  burnTime?: string;
  waxType?: string;
  location?: string;
  warnings?: string;
  sku?: string;
  isActive?: boolean;
  labelTemplateId?: number;
}