# lib/api-zod/src/generated/types/labelTemplate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Label Studio API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LabelTemplateZones } from "./labelTemplateZones";

export interface LabelTemplate {
  id: number;
  name: string;
  description?: string | null;
  labelSheetId?: number | null;
  zones: LabelTemplateZones;
  previewNotes?: string | null;
  createdAt: Date;
  updatedAt: Date;
}