# lib/api-zod/src/generated/types/printJob.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Label Studio API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PrintJobItem } from "./printJobItem";
import type { PrintJobStatus } from "./printJobStatus";

export interface PrintJob {
  id: number;
  name: string;
  labelSheetId: number;
  labelSheetName: string;
  labelSheetCode: string;
  labelsPerSheet: number;
  items: PrintJobItem[];
  totalLabels: number;
  totalSheets: number;
  status: PrintJobStatus;
  notes?: string | null;
  createdAt: Date;
  updatedAt: Date;
}