# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Label Studio API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProductProductType } from "./productProductType";

export interface Product {
  id: number;
  productType: ProductProductType;
  name: string;
  scentName: string;
  scentNotes?: string | null;
  size: string;
  weight?: string | null;
  ingredients?: string | null;
  instructions?: string | null;
  burnTime?: string | null;
  waxType?: string | null;
  location?: string | null;
  warnings?: string | null;
  sku?: string | null;
  isActive: boolean;
  labelTemplateId?: number | null;
  createdAt: Date;
  updatedAt: Date;
}