# lib/api-zod/src/generated/types/updateLabelTemplateBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Label Studio API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateLabelTemplateBodyZones } from "./updateLabelTemplateBodyZones";

export interface UpdateLabelTemplateBody {
  name?: string;
  description?: string;
  labelSheetId?: number | null;
  zones?: UpdateLabelTemplateBodyZones;
  previewNotes?: string;
}